+++ tests/unit_tests/test_mpl_themes.py
@@ -23,10 +23,3 @@ def test_font_scale():
     vizta.mpl.set_theme(font_scale=2)
     with pytest.raises(ValueError):
         vizta.mpl.set_theme(font_scale="blah")
-
-
-def test_plotly_styles():
-    """Really just check that theme doesn't error"""
-    vizta.mpl.plotly_vizta_theme()
-    with pytest.raises(ValueError):
-        vizta.mpl.set_theme(style="blah")
