+++ vizta/__init__.py
@@ -1,5 +1,5 @@
 """See the README for detailed documentation and examples."""
 from .version import _get_version
-from . import mpl
+from . import mpl, plotly
 
 __version__ = _get_version()
