+++ vizta/mpl/__init__.py
@@ -1,2 +1,1 @@
 from .themes import set_theme
-from .themes import plotly_vizta_theme
