+++ vizta/mpl/themes.py
@@ -7,9 +7,6 @@ from typing import List, Tuple, Dict, Any, Optional, Union
 
 import seaborn as sns
 
-import plotly.graph_objects as go
-import plotly.io as pio
-
 from .. import palettes  # Needed to edit seaborn palette dict.
 
 
@@ -56,27 +53,6 @@ def set_theme(
     )
 
     return theme.primary, theme.accent
-
-
-def plotly_vizta_theme():
-    pio.templates["talusbio"] = go.layout.Template(
-        layout={
-            "title": {
-                "font": {
-                    "family": "Helvetica,Arial,sans-serif",
-                    "size": 30,
-                    "color": "#333",
-                }
-            },
-            "font": {
-                "family": "Helvetica,Arial,sans-serif",
-                "size": 16,
-                "color": "#333",
-            },
-            # Colorways
-            "colorway": ["#308AAD", "#C8102E", "#96D8D8", "#B2EE79"],
-        }
-    )
 
 
 @dataclass
